+++ lib/src/decoder.dart
@@ -1,5 +1,6 @@
 library toml.src.decoder;
 
+export 'decoder/exception/invalid_date_time.dart';
 export 'decoder/exception/invalid_escape_sequence.dart';
 export 'decoder/exception/not_a_table.dart';
 export 'decoder/exception/parser.dart';